Arrange classes into packages
Commands and CommandResults classes moved to com.duke.command package.

DukeException classes moved to com.duke.exception package.

Task classes moved to com.duke.tasks package.

Parser, Ui, Tasklist, Storage moved to com.duke.modules package,
