+++ src/main/java/Duke.java
@@ -1,3 +1,8 @@
+import com.duke.modules.Parser;
+import com.duke.modules.Storage;
+import com.duke.modules.TaskList;
+import com.duke.modules.Ui;
+
 public class Duke {
     private Parser cmdParser;
