+++ src/main/java/com/duke/command/Command.java
@@ -1,3 +1,7 @@
+package com.duke.command;
+
+import com.duke.modules.TaskList;
+
 public abstract class Command {
   protected TaskList taskList;
