+++ src/main/java/com/duke/command/CommandResult.java
@@ -1,3 +1,5 @@
+package com.duke.command;
+
 public class CommandResult{
   private boolean isShutdown = false;
   private String resultMessage;
