+++ src/main/java/com/duke/exception/DukeEmptyListException.java
@@ -1,3 +1,5 @@
+package com.duke.exception;
+
 public class DukeEmptyListException extends DukeException {
   public DukeEmptyListException(String errormessage) {
     super(errormessage);
