+++ src/main/java/com/duke/exception/DukeException.java
@@ -1,3 +1,5 @@
+package com.duke.exception;
+
 public class DukeException extends Exception {
   public DukeException(String errormessage) {
     super(errormessage);
