+++ src/main/java/com/duke/exception/DukeFileParseException.java
@@ -1,3 +1,5 @@
+package com.duke.exception;
+
 public class DukeFileParseException extends DukeException {
   public DukeFileParseException(String errormessage) {
     super(errormessage);
