+++ src/main/java/com/duke/exception/DukeInvalidArgumentException.java
@@ -1,3 +1,5 @@
+package com.duke.exception;
+
 public class DukeInvalidArgumentException extends DukeException{
   public DukeInvalidArgumentException(String errormessage) {
     super(errormessage);
