+++ src/main/java/com/duke/modules/TaskList.java
@@ -1,10 +1,9 @@
+package com.duke.modules;
+
+import com.duke.tasks.Task;
+
 import java.io.IOException;
-import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
-import java.time.format.DateTimeParseException;
 import java.util.ArrayList;
-import java.util.Date;
-import java.util.zip.DataFormatException;
 
 public class TaskList {
   private Storage storage;
