+++ src/main/java/com/duke/tasks/Deadline.java
@@ -1,3 +1,5 @@
+package com.duke.tasks;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
