+++ src/main/java/com/duke/tasks/Todo.java
@@ -1,3 +1,5 @@
+package com.duke.tasks;
+
 public class Todo extends Task {
 
   public Todo(String description) {
